Provide a way to clean up the class cache memory
This cache previously lacked a method of being cleaned up.
I'm not sure how much memory this cache uses, but since we have a lot of
classes in our code base, I'd rest easier if it could be manually
cleared.
One could potentially implement that the cache is cleared automatically
after a certain amount of time has passed, but this does the job for the
time being.